// src/app/student/store/student.type.ts
export enum StudentActionType{
    STUDENT_LIST = '[Student Page] list of students',

    STUDENT_ADD = '[Student Page] add new student',
    STUDENT_ADD_SUCCESS = '[Student Page] add new student success',

    STUDENT_UPDATE = '[Student Page] update student',
    STUDENT_UPDATE_SUCCESS = '[Student Page] update student sucess',

    STUDENT_DELETE = '[Student Page] delete student',
    STUDENT_DELETE_SUCCESS = '[Student Page] delete student success',

    STUDENT_LOAD = '[Student Page] load student',
    STUDENT_LOAD_SUCCESS = '[Student Page] load student success',
}